refactor(store,cli): home the trace primitives where the ratchet points
Move execution_events (+ its test) into receipts.rs beside the other
per-execution readers, and fold the run-path trace glue into
trace::{episode_tag, capture_or_warn}. Residual growth — the run-path
call sites and one struct-literal field — takes the sanctioned baseline
update.

Refs #1042
