+++ scripts/file-size-baseline.txt
@@ -15,8 +15,8 @@
 1887 bench/terminal_bench_analysis/tb21_evidence_contract.py
 4173 bench/terminal_bench_analysis/tests/test_tb21_analysis.py
 1615 bench/terminal_bench_analysis/tests/test_tb21_evidence_contract.py
-1789 stella-cli/src/agent_tests.rs
-2299 stella-cli/src/agent.rs
+1790 stella-cli/src/agent_tests.rs
+2315 stella-cli/src/agent.rs
 1576 stella-cli/src/candidate_ws.rs
 4389 stella-cli/src/command_deck.rs
 2095 stella-core/src/bus.rs
@@ -30,7 +30,7 @@
 3064 stella-pipeline/src/pipeline.rs
 2360 stella-pipeline/src/pipeline/tests.rs
 2744 stella-protocol/src/event.rs
-2055 stella-store/src/lib.rs
+2056 stella-store/src/lib.rs
 2204 stella-store/src/tests.rs
 1892 stella-store/src/usage.rs
 1566 stella-tools/src/media.rs
